refactor(master): add apply drivers add run cmd and vendor
