+++ pkg/apply/applydrivers/apply_drivers_default.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package apply_drivers
+package applydrivers
 
 import (
 	"fmt"
